train/scripts/desktop_recapture_detect/requirements.txt
# desktop_recapture_detect 训练依赖
# EfficientNet-B0 二分类：真人肖像 vs 桌面翻拍
torch>=2.0.0
torchvision>=0.15.0
onnx>=1.15.0
onnxruntime>=1.18.0
Pillow>=10.0.0
scikit-learn>=1.4.0
tqdm>=4.66.0
PyYAML>=6.0.1
numpy>=1.24.0
